src/backend/requirements.txt
#for imports
azure-functions
langchain
langchain-core
langchain-ollama
httpx
requests
pymongo
python-dotenv